Refactoring edge class for efficiency and ease of use
Started by redoing the constructor.  It now expects an additional parameter bezierPoint = {x:,y:}.  Naming of variables has been improved for readability, these changes have been propogated to the rest of the code base.  Member variables have been increased to accommodate more necessary information, like start and end point, and the type of cost (weight vs grammar) the edge is currently using.  Code has been included in the constructor to calculate the costType as well as the bezier point is not provided, as only two arguments are required:  the starting node and destination node.

Optimized drawing, now the start and end points are calculated only when they change instead of every call to draw.  For now this is only in the constructor, need to ensure this occurs when nodes are moved as well.