Updated edge class and implemented dashed and solid edge subclasses
Repleted the edge class with some getter, setter, and helper functions to make interfacing with it more pleasant.  Please read the comments in the file before using it, bypassing these methods and directly accessing data members is not advised.

Implemented two subclasses and plugged them into the toolbar via the edge-tool class.  The Edge Class is now an abstract superclass not intended for direct use.