Created infrastructure for Sidebar Context Switching and Updating
app.js and ui.js were updated to pass a graph to sidebar.js

tools now have a sidebarType, a string, to refer to their associated sidebar context

toolbar.js modified to make a call for changing sidebar context when a different tool is selected

mouse-handler.js altered to update the content of the sidebar information when the mouse is let up

sidebar.js is now responsible for outer container and is the parent of the inner container code:  sidebar-content is the superclass of the inner container code files.  Currently there are only two inner container class files, sidebar-edge and sidebark-node which essentially serve as examples at this point and will be replaced. The basic premise is that switching tools will switch the inner container of the sidebar to present relevant contextual information.  See the sidebar.js documentation for more info.