Created branch and started infrastructure for emulating algorithms
stepper will act as interface for user to control flow of emulation, algorithm will act as an abstract class to implement the logic of the specific algorithms, graphLogic and graphVisuals will contain libraries with functionality common to all or most algorithms.  queue is a wrapper implementation that is more efficient for larger arrays sizes. see notes in file for more details