Redesigned algorithm superclass
Before the idea was to keep track of both the logical state of the algorithm emulation as well as the graphical state and to step back and forth inbetween them.  After a little thought I realized this was quite wasteful, and now the algorithm is run through completion once and the graphical changes induced by this are stored in an array where they can be stepped through at will.  For more details and a forward facing API, see the comments in the file.