Step details added to GUI
Had to put the code for changing the html of the elements created in the sidebar in stepper.  Feels hacky but anywhere lower makes no sense and any higher is way too much work.  Should refactor that whole chain of dependencies and compress them, ask Allen if he had other plans for that particular structure.