Algorithm infrastructure refactor
Things got pretty messy but this was a big step in the right direction.  ui/algorithm.js had a lot of excessive code and served almost entirely as a useless wrapper/go-between.  It has been removed and sidebar-algorithm now interfaces directly with the stepper.  Some lint issues were fixed, and some function extractions were made in sidebar-algorithm for readability.  Much more work still to be done.